Fix clippy lints surfaced by newer toolchain
CI's clippy (1.96) flags lints my local 1.91 missed: redundant casts,
sort_by -> sort_by_key, a useless into_iter, and manual checked division.
All behavior-preserving; fmt/clippy/tests green on 1.96.

+++ src/eval.rs
@@ -237,8 +237,8 @@ impl PestoEval {
             if king_sq < 64 {
                 let shield_zone_mask = get_king_shield_zone_mask(color, king_sq);
                 let shield_pawns = popcnt(shield_zone_mask & friendly_pawns);
-                mg[color] += shield_pawns as i32 * self.weights.king_safety_pawn_shield_bonus[0];
-                eg[color] += shield_pawns as i32 * self.weights.king_safety_pawn_shield_bonus[1];
+                mg[color] += shield_pawns * self.weights.king_safety_pawn_shield_bonus[0];
+                eg[color] += shield_pawns * self.weights.king_safety_pawn_shield_bonus[1];
 
                 let enemy_king_sq = board.pieces[enemy_color][KING].trailing_zeros() as usize;
                 if enemy_king_sq < 64 {
